scripts/python/requirements.txt
pillow==12.1.1
pillow_heif==1.2.0
